refactor(frames): styleFrameClipOctagon and styleFrameClipKranox
