--- packages/frames/src/createFrameKranoxClip/index.ts
@@ -1,1 +1,0 @@
-export * from './createFrameKranoxClip.js'
