--- packages/frames/src/createFrameOctagonClip/index.ts
@@ -1,1 +1,0 @@
-export * from './createFrameOctagonClip.js'
