+++ packages/frames/src/index.ts
@@ -1,5 +1,5 @@
 export * from './types.js'
 export * from './formatFrameSVGPath/index.js'
 export * from './renderFrameSVGPaths/index.js'
-export * from './createFrameOctagonClip/index.js'
-export * from './createFrameKranoxClip/index.js'
+export * from './styleFrameClipOctagon/index.js'
+export * from './styleFrameClipKranox/index.js'
