+++ packages/frames/src/styleFrameClipKranox/index.ts
@@ -1,0 +1,1 @@
+export * from './styleFrameClipKranox.js'
