+++ packages/frames/src/styleFrameClipKranox/styleFrameClipKranox.basic.sandbox.tsx
@@ -1,4 +1,4 @@
-import { createFrameKranoxClip } from '@arwes/frames'
+import { styleFrameClipKranox } from '@arwes/frames'
 
 const root = document.querySelector('#root')!
 const box = document.createElement('div')
@@ -8,5 +8,5 @@ Object.assign(box.style, {
   width: '200px',
   height: '400px',
   background: '#077',
-  clipPath: createFrameKranoxClip({ squareSize: '1rem' })
+  clipPath: styleFrameClipKranox({ squareSize: '1rem' })
 })
