+++ packages/frames/src/styleFrameClipKranox/styleFrameClipKranox.lengths.sandbox.tsx
@@ -1,4 +1,4 @@
-import { createFrameKranoxClip } from '@arwes/frames'
+import { styleFrameClipKranox } from '@arwes/frames'
 
 const root = document.querySelector('#root')!
 const box = document.createElement('div')
@@ -8,7 +8,7 @@ Object.assign(box.style, {
   width: '200px',
   height: '400px',
   background: '#077',
-  clipPath: createFrameKranoxClip({
+  clipPath: styleFrameClipKranox({
     padding: 4,
     strokeWidth: 2,
     squareSize: 12,
