+++ packages/frames/src/styleFrameClipOctagon/index.ts
@@ -1,0 +1,1 @@
+export * from './styleFrameClipOctagon.js'
