+++ packages/frames/src/styleFrameClipOctagon/styleFrameClipOctagon.basic.sandbox.tsx
