+++ packages/frames/src/styleFrameClipOctagon/styleFrameClipOctagon.corners.sandbox.tsx
