+++ packages/frames/src/styleFrameClipOctagon/styleFrameClipOctagon.test.ts
